2009-03-26  Sancho Lerena <slerena@artica.es>
	* pandoradb.sql: SLA parameters now accept float values.
	
	* extras/pandoradb_migrate_v2.x_to_v3.0.sql: SLA parameters now accept 
	float values after a table update.
	
	* godmode/setup/filemgr.php: New feature: a file manager, accesible only by
	pandora Administrators. Have some security checks but is not intended to be
	used by anyone not admin.
	
	* godmode/menu.php: Added filemanager and fixed minor problem under SNMP
	Godmode (fake submenu).



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@1572 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/pandoradb.sql
@@ -716,9 +716,9 @@ CREATE TABLE IF NOT EXISTS `treport_content_sla_combined` (
   `id` INTEGER UNSIGNED NOT NULL auto_increment,
   `id_report_content` INTEGER UNSIGNED NOT NULL,
   `id_agent_module` int(10) unsigned NOT NULL,
-  `sla_max` int(11) NOT NULL default 0,
-  `sla_min` int(11) NOT NULL default 0,
-  `sla_limit` int(11) NOT NULL default 0,
+  `sla_max` double(18,2) NOT NULL default 0,
+  `sla_min` double(18,2) NOT NULL default 0,
+  `sla_limit` double(18,2) NOT NULL default 0,  
   PRIMARY KEY(`id`),
   FOREIGN KEY (`id_report_content`) REFERENCES treport_content(`id_rc`)
      ON UPDATE CASCADE ON DELETE CASCADE,
